requirements.txt
gymnasium==0.29.1
numpy==1.26.4
opencv-python==4.10.0
git+https://github.com/scikit-video/scikit-video.git
termcolor
scipy
wandb==0.17.8
dm_control==1.0.20
scikit-image==0.24.0
hydra-core==1.3.2
mujoco==3.1.6